+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix: [#798] Setting waypoints on multitile track/road elements corrupts the position.
 - Fix: [#801] Initial save path does not contain a trailing slash.
 - Fix: [#807] Incorrect vehicle animation for speed based animations like hydrofoils when at max speed.
+- Change: [#787] Scenery and building interaction is now disabled when see-through.
 
 21.02 (2021-02-20)
 ------------------------------------------------------------------------